Move currency_emojis.py to "shared"
