+++ handlers/currencies/__init__.py
@@ -1,5 +1,4 @@
 from .convert_currency import *
-from .currency_emojis import *
 from .currency_triggers import *
 from .delete_currency_message import *
 from .get_rates import *
