--- handlers/exchange/currency_emojis.py
@@ -1,1 +1,0 @@
-../currencies/currency_emojis.py
