+++ shared/currency_emojis.py
